rename plugin examples [skip-ci]
